fix: story sorting in sidebar of storybook, template generator fixed as well
