+++ src/components/templates/MSWTemplates/fetch/App.stories.js
@@ -3,7 +3,7 @@ import { rest } from 'msw';
 import { App } from './App';
 
 export default {
-  title: 'Examples/MSW Demos/Axios',
+  title: 'Templates & Guides/MSW Demos/Fetch',
   component: App,
 };
