+++ src/components/templates/MSWTemplates/fetch/App.test.js
