+++ src/components/templates/MSWTemplates/react-router-react-query/pages/Home.js
