+++ src/components/templates/MSWTemplates/react-router-react-query/pages/character/Character.js
