+++ src/components/templates/NewComponentTemplate/NewComponentTemplate.styles.ts
