+++ src/components/templates/NewComponentTemplate/NewComponentTemplate.test.tsx
