+++ src/components/templates/NewComponentTemplate/NewComponentTemplate.tsx
