+++ src/components/templates/NewComponentTemplate/README.md
