+++ src/components/templates/NewComponentTemplate/index.ts
