+++ src/stories/Introduction.stories.mdx
@@ -1,6 +1,6 @@
 import { Meta } from '@storybook/addon-docs';
 
-<Meta title="Guides/Storybook Introduction" />
+<Meta title="Templates & Guides/Storybook Introduction" />
 
 <style>{`
   .subheading {
