+++ src/stories/msw-introduction.stories.mdx
@@ -1,7 +1,7 @@
 import { Meta } from '@storybook/addon-docs/blocks'
 import LinkTo from '@storybook/addon-links/react'
 
-<Meta title="Guides/MSW Introduction" />
+<Meta title="Templates & Guides/MSW Introduction" />
 
 <header className="introduction-header">
   <img src="https://msw-sb.vercel.app/logo.png" width="200" height="172" />
